# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Click Proposta v1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/proposal": {
      "post": {
        "operationId": "createProposal",
        "tags": [
          "Proposals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "urlLogoImage": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uri"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "customersId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                  },
                  "welcomeDescription": {
                    "type": "string",
                    "minLength": 1
                  },
                  "whyUs": {
                    "type": "string",
                    "minLength": 1
                  },
                  "challenge": {
                    "type": "string",
                    "minLength": 1
                  },
                  "solution": {
                    "type": "string",
                    "minLength": 1
                  },
                  "services": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "price": {
                          "type": "number",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "servicesId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                        }
                      },
                      "required": [
                        "price",
                        "servicesId"
                      ]
                    }
                  },
                  "results": {
                    "type": "string",
                    "minLength": 1
                  },
                  "discount": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  }
                },
                "required": [
                  "urlLogoImage",
                  "title",
                  "customersId",
                  "welcomeDescription",
                  "whyUs",
                  "challenge",
                  "solution",
                  "services",
                  "results",
                  "discount"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/proposal/confirm-sending/{proposalId}": {
      "patch": {
        "operationId": "confirmSendingProposal",
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "proposalId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "default": 204,
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/proposal/approve-proposal/{proposalId}": {
      "patch": {
        "operationId": "approveProposal",
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "proposalId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "default": 204,
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/payment": {
      "post": {
        "operationId": "createNewPayment",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "price": {
                    "default": 2499,
                    "type": "number"
                  },
                  "quantity": {
                    "default": 1,
                    "type": "number"
                  },
                  "customer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 3
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "cellphone": {
                        "type": "string",
                        "minLength": 10,
                        "maxLength": 11
                      },
                      "cpf": {
                        "type": "string",
                        "minLength": 11,
                        "maxLength": 14
                      }
                    },
                    "required": [
                      "name",
                      "email",
                      "cellphone",
                      "cpf"
                    ]
                  }
                },
                "required": [
                  "customer"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/draft-proposals/{customerId}/last": {
      "get": {
        "operationId": "getLastDraftProposal",
        "tags": [
          "DraftProposals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "customerId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "proposalDraft": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "welcomeDescription": {
                          "type": "string"
                        },
                        "whyUs": {
                          "type": "string"
                        },
                        "challenge": {
                          "type": "string"
                        },
                        "solution": {
                          "type": "string"
                        },
                        "results": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "title",
                        "welcomeDescription",
                        "whyUs",
                        "challenge",
                        "solution",
                        "results"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "proposalDraft"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}